==> 5-VAR.Rmd <==
---
title: "VAR Models"
author: "Sarthak Saluja"
output: html_document
---

Purpose :
1. Clean matched topic data for VAR models
2. Produce nice timeseries plots
3. Run and tune VAR models
4. Perform Granger Causality Test
5. Compute Impulsive Response Functions

Load libraries :

```{r, message = F}
library(xts)
library(vars)
library(bruceR)
library(ggplot2)
library(dplyr)
library(tidyverse)
library(stringr)
library(tidyr)
library(tseries)

library(devtools)
source_url("https://raw.githubusercontent.com/anguyen1210/var-tools/master/R/extract_varirf.R")
```

Load data :

```{r}
rm(list = ls())

suppress <- read.csv('topics/suppress.csv')
singhu <- read.csv('topics/singhu.csv')
rihanna <- read.csv('topics/rihanna.csv')
republic_day <- read.csv('topics/republic day.csv')
crony <- read.csv('topics/crony capitalism.csv')
media <- read.csv('topics/media hypocracy.csv')
covid <- read.csv('topics/covid.csv')
```

Data Cleaning :

```{r}
all_topics <- list(suppress, singhu, rihanna, republic_day, crony, media, covid)


all_topics_cleaned <- lapply(all_topics, function(x){

  x <- x |>
    filter(!is.na(gamma_core) & !is.na(gamma_periphery) &
             gamma_core != 0 & gamma_periphery != 0) |>
    mutate(document = as.Date(document)) |>
    filter(document > as.Date('2020-11-01')) |> # Filter values due to missing data
    mutate(log_gamma_core = log(gamma_core)/(1-log(gamma_core)),
           log_gamma_periphery = log(gamma_periphery)/(1-log(gamma_periphery))) # Log to account for skewness

  x <- xts(x[, 6:7], order.by = x$document) # Create time series object
  return(x)
})

```

Check for Stationarity :

```{r}
lapply(all_topics_cleaned, function(x){
  print(adf.test(x$log_gamma_core))
  print(adf.test(x$log_gamma_periphery))
})

```

Plot Time Series :

```{r, out.width=5, out.height=8}

title <- c('Suppress Dissent',
            'Singhu Border',
            'Rihanna',
            'Republic Day Violence',
            'Crony Capitalism',
            'Media Hypocricy',
            'Covid')

RD = F
Rihanna = F
legend <- c('Periphery' = '#46A6C3', 'Core' = '#1E4E5C')

lapply(1:length(all_topics_cleaned), function(x){

  df <- all_topics_cleaned[[x]]
  name <- title[[x]]

  if(name == 'Republic Day Violence'){
    RD = T
  }

  if(name == 'Rihanna'){
    Rihanna = T
  }

  p <- df |>
    as.data.frame() |>
    rownames_to_column('document') |>
    mutate(document = as.Date(document)) |>
    ggplot() +
    geom_line(aes(x = document, y = log_gamma_periphery, color = 'Periphery')) +
    geom_line(aes(x = document, y = log_gamma_core, color = 'Core')) +
    {if(RD) list(
    geom_vline(aes(xintercept = as.Date('2021-01-26')),
                   colour = 'red',
                   linetype = 'dashed',
                   alpha = 0.5),
    geom_segment(aes(x = as.Date('2021-02-15'),
                     y = -10,
                     xend = as.Date('2021-01-30'),
                     yend = -10),
                 arrow = arrow(length = unit(0.2, "cm")),
                 colour = 'red'),
    annotate('text',
             x = as.Date('2021-03-20'),
             y = -10,
             label = 'Republic Day',
             colour = 'red'))} +
    {if(Rihanna) list(
    geom_vline(aes(xintercept = as.Date('2021-02-02')),
                   colour = 'red',
                   linetype = 'dashed',
                   alpha = 0.5),
    geom_segment(aes(x = as.Date('2021-02-20'),
                     y = -9.7,
                     xend = as.Date('2021-02-06'),
                     yend = -9.7),
                 arrow = arrow(length = unit(0.2, "cm")),
                 colour = 'red'),
    annotate('text',
             x = as.Date('2021-04-01'),
             y = -9.7,
             label = "Rihanna's Tweet",
             colour = 'red'))} +
    labs(title = paste0('Topic : ', name),
         x = 'date',
         y = 'topic prevelance',
         color = 'Legend') +
    scale_color_manual(values = legend) +
    theme_bw() +
    scale_x_date(date_labels = "%b-%Y",
                 date_breaks = '1 month') +
    theme(axis.text.x = element_text(size = 10, angle = 90),
          legend.title = element_blank())

  ggsave(paste0('plots/Time series/', name, '.png'),
         p,
         width = 7,
         height = 3)
})
```


```{r}
options(scipen = 99)

# Evaluate metrics

VAR_eval <- lapply(1 : length(all_topics_cleaned), function(x){

  df <- all_topics_cleaned[[x]]
  name <- title[[x]]

  df <- data.frame(VARselect(df,
                             lag.max = 15,
                             type = 'const')$selection)

  return(df)
})

VAR_eval <- do.call(cbind, VAR_eval)
colnames(VAR_eval) <- title

VAR_eval <- VAR_eval |>
  rownames_to_column('metric')

# Find the optimal lags

VAR_eval_optimal <- VAR_eval |>
  filter(metric == 'AIC(n)') |>
  t() |>
  as.data.frame() |>
  rownames_to_column('topic') |>
  filter(topic != 'metric') |>
  select(topic, AIC = V1)

VAR_eval_optimal

write.csv(VAR_eval_optimal,
          'csv/VAR_selection.csv',
          row.names = F)
```

```{r}

# Fit models


suppress.VAR <- VAR(all_topics_cleaned[[1]],
                    p = as.numeric(VAR_eval_optimal$AIC[1]),
                    type = 'both')

singhu.VAR <- VAR(all_topics_cleaned[[2]],
                    p = as.numeric(VAR_eval_optimal$AIC[2]),
                    type = 'both')

rihanna.VAR <- VAR(all_topics_cleaned[[3]],
                    p = as.numeric(VAR_eval_optimal$AIC[3]),
                    type = 'both')

republic.VAR <- VAR(all_topics_cleaned[[4]],
                    p = as.numeric(VAR_eval_optimal$AIC[4]),
                    type = 'both')

crony.VAR <- VAR(all_topics_cleaned[[5]],
                    p = as.numeric(VAR_eval_optimal$AIC[5]),
                    type = 'both')

media.VAR <- VAR(all_topics_cleaned[[6]],
                    p = as.numeric(VAR_eval_optimal$AIC[6]),
                    type = 'both')

covid.VAR <- VAR(all_topics_cleaned[[7]],
                    p = as.numeric(VAR_eval_optimal$AIC[7]),
                    type = 'both')


# Plot IRF

VAR <- list(suppress.VAR, singhu.VAR, rihanna.VAR, republic.VAR, crony.VAR, media.VAR, covid.VAR)

irfc2p <-
  lapply(VAR,
         irf,
         impulse = 'log_gamma_core',
         response = 'log_gamma_periphery',
         n.ahead = 15,
         ortho = F,
         runs = 100,
         seed = 1234)

irfp2c <-
  lapply(VAR,
         irf,
         impulse = 'log_gamma_periphery',
         response = 'log_gamma_core',
         n.ahead = 15,
         ortho = F,
         runs = 100,
         seed = 12345)

cirfc2p <- lapply(VAR,
         irf,
         impulse = 'log_gamma_core',
         response = 'log_gamma_periphery',
         n.ahead = 15,
         ortho = F,
         runs = 100,
         seed = 1,
         cumulative = T)

cirfp2c <-
  lapply(VAR,
         irf,
         impulse = 'log_gamma_periphery',
         response = 'log_gamma_core',
         n.ahead = 15,
         ortho = F,
         runs = 100,
         seed = 11,
         cumulative = T)
```


```{r}
create_irf <- function(start = 'core',
                       cumulative = F){

  if (start == 'core' & cumulative != T){

    irf <- irfc2p
    main = 'Impulsive Response from Core'
    y = 'logit gamma periphery'
    filename = '_core_to_periphery'

  } else if (start == 'core' & cumulative == T){

    irf <- cirfc2p
    main = 'Cumulative Impulsive Response from Core'
    y = 'logit gamma periphery'
    filename = '_cumulative_core_to_periphery'

  } else if (start == 'periphery' & cumulative != T){

    irf <- irfp2c
    main = 'Impulsive Response from Periphery'
    filename = '_periphery_to_core'
    y = 'logit gamma core'

  } else if (start == 'periphery' & cumulative == T){

    irf <- cirfp2c
    main = 'Cumulative Impulsive Response from Periphery'
    filename = '_cumulative_periphery_to_core'
    y = 'logit gamma core'
  }

  asdf <- lapply(1:length(irf), function(x){

    to_clean <- irf[[x]]

    df <- extract_varirf(to_clean)
    title <- title[[x]]

    df <- df |>
      select(-period)

    colnames(df) <- c(paste0('main_', title),
                      paste0('lower_', title),
                      paste0('upper_', title))
  return(df)})

  asdf <- do.call(cbind, asdf)
  asdf$day <- 0:15

  asdf <- asdf |>
  pivot_longer(!day,
               names_to = c('type', 'topic'),
               names_sep = '_',
               values_to = 'values')

  p <- asdf |>
  ggplot(aes(x = day,
                y = values,
                group = type)) +
  geom_line(aes(linetype = type,
                color = type)) +
  facet_wrap(~topic, ncol = 4) +
    {if(cumulative == F)scale_y_continuous(breaks = c(-0.25, 0, 0.25, 0.5),
                     limits = c(-0.25, 0.5))} +
        {if(cumulative == T)scale_y_continuous(breaks = c(-0.5, 0, 0.5, 1, 1.5, 2),
                     limits = c(-0.5, 2))} +
  scale_linetype_manual(values = c('dashed', 'solid', 'dashed')) +
  scale_color_manual(values = c('#46A6C3', '#1E4E5C', '#46A6C3')) +
  geom_hline(aes(yintercept = 0)) +
  theme_bw() +
  ylab(y) +
  labs(title = main,
       subtitle = 'Dashed lines represent the 95% CI') +
  theme(legend.position = 'none')

  save_at <- paste0('plots/VAR/irf', filename, '.png')
  print(paste0('Saving to : ', save_at))

  ggsave(save_at, p)

  return(p)

}

create_irf(start = 'core',
           cumulative = T)

create_irf(start = 'core',
           cumulative = F)

create_irf(start = 'periphery',
           cumulative = T)

create_irf(start = 'periphery',
           cumulative = F)


```

Granger Causality :

```{r}

lapply(1 : length(VAR), function(x){
  var <- VAR[[x]]
  name <- title[x]

  granger_causality(var,
       file = paste0('causality/', name, '.doc'))
})

```
